Move Groups tests into specific folders
