+++ src/containers/Groups/ContactsFilter/__tests__/ContactsFilter.js
@@ -1,25 +1,20 @@
 import React from 'react';
 
-import SurveyContactsFilter from '../SurveyContactsFilter';
+import ContactsFilter from '..';
+
 import {
   createMockStore,
   renderShallow,
   testSnapshotShallow,
   createMockNavState,
-} from '../../../../testUtils';
-import { navigatePush } from '../../../actions/navigation';
-import * as filterUtils from '../../../utils/filters';
+} from '../../../../../testUtils';
+import { navigatePush } from '../../../../actions/navigation';
+import * as filterUtils from '../../../../utils/filters';
 
-jest.mock('../../../actions/navigation', () => ({
+jest.mock('../../../../actions/navigation', () => ({
   navigatePush: jest.fn(() => ({ type: 'test' })),
 }));
-jest.mock('../../../actions/surveys', () => ({
-  getSurveyQuestions: jest.fn(() => ({
-    type: 'surveyQuestions',
-    response: [{ id: '1' }, { id: '2' }],
-  })),
-}));
-jest.mock('../../../actions/labels', () => ({
+jest.mock('../../../../actions/labels', () => ({
   getOrgLabels: jest.fn(() => ({
     type: 'orgLabels',
     response: [{ id: '3' }, { id: '4' }],
@@ -28,6 +23,7 @@ jest.mock('../../../actions/labels', () => ({
 
 const store = createMockStore({});
 const timeFilter30 = { id: 'time30', value: 30, text: 'Last 30 days' };
+const organization = { id: '1' };
 const filters = {
   unassigned: {
     id: 'unassigned',
@@ -36,17 +32,14 @@ const filters = {
   },
   time: timeFilter30,
 };
-const survey = { id: '11' };
-const organization = { id: '22' };
 
-describe('SurveyContactsFilter', () => {
+describe('ContactsFilter', () => {
   const onFilter = jest.fn();
   const component = (
-    <SurveyContactsFilter
+    <ContactsFilter
       navigation={createMockNavState({
         onFilter,
         filters,
-        survey,
         organization,
       })}
     />
